updated clerk example to app router
