--- examples/with-clerk/.env.local.example
@@ -1,3 +1,0 @@
-NEXT_PUBLIC_CLERK_FRONTEND_API=your_clerk_frontend_api
-CLERK_API_KEY=your_clerk_api_key
-CLERK_JWT_KEY=your_clerk_jwt_key
