+++ examples/with-clerk/app/layout.tsx
@@ -1,0 +1,58 @@
+import '../styles/globals.css'
+import { ClerkProvider, SignedIn, SignedOut, UserButton } from '@clerk/nextjs'
+import { Inter } from 'next/font/google'
+import Image from 'next/image'
+import Script from 'next/script'
+import styles from '../styles/Header.module.css'
+import Link from 'next/link'
+
+const inter = Inter({ subsets: ['latin'] })
+
+export const metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+}
+
+const Header = () => (
+  <header className={styles.header}>
+    <div className={styles.left}>
+      <Link href="/" className={styles.logo}>
+        <Image src="/logo.svg" width="32" height="32" alt="Logo" />
+        <span className={styles.appName}>Your application</span>
+      </Link>
+    </div>
+    <div className={styles.right}>
+      <SignedOut>
+        <Link href="/sign-in">Sign in</Link>
+      </SignedOut>
+      <SignedIn>
+        <UserButton />
+      </SignedIn>
+    </div>
+  </header>
+)
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <ClerkProvider>
+      <html lang="en">
+        <body className={inter.className}>
+        <Header />
+    <main>{children}</main>
+        </body>
+        <Script
+          src="https://cdn.jsdelivr.net/npm/prismjs@1/components/prism-core.min.js"
+          strategy="beforeInteractive"
+        />
+        <Script
+          src="https://cdn.jsdelivr.net/npm/prismjs@1/plugins/autoloader/prism-autoloader.min.js"
+          strategy="beforeInteractive"
+        />
+      </html>
+    </ClerkProvider>
+  )
+}
