+++ examples/with-clerk/app/page.tsx
@@ -1,8 +1,8 @@
-import React from 'react'
-import Head from 'next/head'
-import Link from 'next/link'
+import { SignedIn, SignedOut } from "@clerk/nextjs";
 import styles from '../styles/Home.module.css'
-import { SignedIn, SignedOut } from '@clerk/nextjs'
+import Link from "next/link";
+import Head from "next/head";
+import { APIRequest } from "./api-request";
 
 const ClerkFeatures = () => (
   <Link href="/user" className={styles.cardContent}>
@@ -35,16 +35,6 @@ const SignupLink = () => (
     </div>
   </Link>
 )
-
-const apiSample = `import { getAuth } from '@clerk/nextjs/server'
-
-export default function handler(req, res) {
-  const { sessionId, userId } = getAuth(req);
-  if (!sessionId) {
-    return res.status(401).json({ id: null });
-  }
-  return res.status(200).json({ id: userId });
-}`
 
 // Main component using <SignedIn> & <SignedOut>.
 //
@@ -112,73 +102,6 @@ const Main = () => (
   </main>
 )
 
-const APIRequest = () => {
-  React.useEffect(() => {
-    if (window.Prism) {
-      window.Prism.highlightAll()
-    }
-  })
-  const [response, setResponse] = React.useState(
-    '// Click above to run the request'
-  )
-  const makeRequest = async () => {
-    setResponse('// Loading...')
-
-    try {
-      const res = await fetch('/api/getAuthenticatedUserId')
-      const body = await res.json()
-      setResponse(JSON.stringify(body, null, '  '))
-    } catch (e) {
-      setResponse(
-        '// There was an error with the request. Please contact support@clerk.dev'
-      )
-    }
-  }
-  return (
-    <div className={styles.backend}>
-      <h2>API request example</h2>
-      <div className={styles.card}>
-        <button
-          target="_blank"
-          rel="noreferrer"
-          className={styles.cardContent}
-          onClick={() => makeRequest()}
-        >
-          <img src="/icons/server.svg" />
-          <div>
-            <h3>fetch('/api/getAuthenticatedUserId')</h3>
-            <p>
-              Retrieve the user ID of the signed in user, or null if there is no
-              user
-            </p>
-          </div>
-          <div className={styles.arrow}>
-            <img src="/icons/download.svg" />
-          </div>
-        </button>
-      </div>
-      <h4>
-        Response
-        <em>
-          <SignedIn>
-            You are signed in, so the request will return your user ID
-          </SignedIn>
-          <SignedOut>
-            You are signed out, so the request will return null
-          </SignedOut>
-        </em>
-      </h4>
-      <pre>
-        <code className="language-js">{response}</code>
-      </pre>
-      <h4>pages/api/getAuthenticatedUserId.js</h4>
-      <pre>
-        <code className="language-js">{apiSample}</code>
-      </pre>
-    </div>
-  )
-}
-
 // Footer component
 const Footer = () => (
   <footer className={styles.footer}>
@@ -193,8 +116,9 @@ const Footer = () => (
   </footer>
 )
 
-const Home = () => (
-  <div className={styles.container}>
+export default function Home() {
+  return (
+    <div className={styles.container}>
     <Head>
       <title>Create Next App</title>
       <link rel="icon" href="/favicon.ico" />
@@ -206,6 +130,5 @@ const Home = () => (
     <Main />
     <Footer />
   </div>
-)
-
-export default Home
+  )
+}
