+++ examples/with-clerk/app/sign-in/[[...sign-in]]/page.tsx
@@ -1,0 +1,5 @@
+import { SignIn } from "@clerk/nextjs";
+
+export default function Page() {
+  return <SignIn />;
+}
