+++ examples/with-clerk/app/sign-up/[[...sign-up]]/page.tsx
@@ -1,0 +1,5 @@
+import { SignUp } from "@clerk/nextjs";
+
+export default function Page() {
+  return <SignUp />;
+}
