--- examples/with-clerk/components/Layout.js
@@ -1,10 +1,0 @@
-import Header from './Header'
-
-const Layout = ({ children }) => (
-  <>
-    <Header />
-    <main>{children}</main>
-  </>
-)
-
-export default Layout
