+++ examples/with-clerk/package.json
@@ -6,9 +6,13 @@
     "start": "next start"
   },
   "dependencies": {
-    "@clerk/nextjs": "4.5.1",
-    "next": "latest",
+    "@clerk/nextjs": "^4.17.3",
+    "@types/node": "20.1.3",
+    "@types/react": "18.2.6",
+    "@types/react-dom": "18.2.4",
+    "next": "13.4.2",
     "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "typescript": "5.0.4"
   }
 }
