--- examples/with-clerk/pages/api/getAuthenticatedUserId.js
@@ -1,9 +1,0 @@
-import { getAuth } from '@clerk/nextjs/server'
-
-export default function handler(req, res) {
-  const { sessionId, userId } = getAuth(req)
-  if (!sessionId) {
-    return res.status(401).json({ id: null })
-  }
-  return res.status(200).json({ id: userId })
-}
