--- examples/with-clerk/pages/sign-in/[[...index]].js
@@ -1,7 +1,0 @@
-import { SignIn } from '@clerk/nextjs'
-
-const SignInPage = () => (
-  <SignIn path="/sign-in" routing="path" signUpUrl="/sign-up" />
-)
-
-export default SignInPage
