--- examples/with-clerk/pages/sign-up/[[...index]].js
@@ -1,7 +1,0 @@
-import { SignUp } from '@clerk/nextjs'
-
-const SignUpPage = () => (
-  <SignUp path="/sign-up" routing="path" signInUrl="/sign-in" />
-)
-
-export default SignUpPage
